# Conflicts:
#	.gitignore
#	.settings/org.eclipse.wst.common.project.facet.core.xml

+++ src/main/java/com/skcc/cloudz/zcp/alert/dao/RuleDao.java
@@ -29,6 +29,6 @@ public interface RuleDao {
 	 * @param id
 	 * @return
 	 */
-	Boolean deleteRule(Long id);
+	Boolean deleteRule(int id);
 	
 }
